+++ rpc/namespaces/ethereum/debug/api.go
@@ -2,7 +2,6 @@ package debug
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -14,11 +13,9 @@ import (
 	"time"
 
 	"github.com/davecgh/go-spew/spew"
-	tmrpctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
 
-	"github.com/cosmos/cosmos-sdk/client"
 	stderrors "github.com/pkg/errors"
 
 	"github.com/cosmos/cosmos-sdk/server"
@@ -43,27 +40,22 @@ type HandlerT struct {
 
 // API is the collection of tracing APIs exposed over the private debugging endpoint.
 type API struct {
-	ctx         *server.Context
-	logger      log.Logger
-	backend     backend.EVMBackend
-	clientCtx   client.Context
-	queryClient *rpctypes.QueryClient
-	handler     *HandlerT
+	ctx     *server.Context
+	logger  log.Logger
+	backend backend.EVMBackend
+	handler *HandlerT
 }
 
 // NewAPI creates a new API definition for the tracing methods of the Ethereum service.
 func NewAPI(
 	ctx *server.Context,
 	backend backend.EVMBackend,
-	clientCtx client.Context,
 ) *API {
 	return &API{
-		ctx:         ctx,
-		logger:      ctx.Logger.With("module", "debug"),
-		backend:     backend,
-		clientCtx:   clientCtx,
-		queryClient: rpctypes.NewQueryClient(clientCtx),
-		handler:     new(HandlerT),
+		ctx:     ctx,
+		logger:  ctx.Logger.With("module", "debug"),
+		backend: backend,
+		handler: new(HandlerT),
 	}
 }
 
@@ -71,109 +63,7 @@ func NewAPI(
 // and returns them as a JSON object.
 func (a *API) TraceTransaction(hash common.Hash, config *evmtypes.TraceConfig) (interface{}, error) {
 	a.logger.Debug("debug_traceTransaction", "hash", hash)
-	// Get transaction by hash
-	transaction, err := a.backend.GetTxByEthHash(hash)
-	if err != nil {
-		a.logger.Debug("tx not found", "hash", hash)
-		return nil, err
-	}
-
-	// check if block number is 0
-	if transaction.Height == 0 {
-		return nil, errors.New("genesis is not traceable")
-	}
-
-	blk, err := a.backend.GetTendermintBlockByNumber(rpctypes.BlockNumber(transaction.Height))
-	if err != nil {
-		a.logger.Debug("block not found", "height", transaction.Height)
-		return nil, err
-	}
-
-	parsedTxs, err := rpctypes.ParseTxResult(&transaction.TxResult)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse tx events: %s", hash.Hex())
-	}
-	parsedTx := parsedTxs.GetTxByHash(hash)
-	if parsedTx == nil {
-		return nil, fmt.Errorf("ethereum tx not found in msgs: %s", hash.Hex())
-	}
-
-	// check tx index is not out of bound
-	if uint32(len(blk.Block.Txs)) < transaction.Index {
-		a.logger.Debug("tx index out of bounds", "index", transaction.Index, "hash", hash.String(), "height", blk.Block.Height)
-		return nil, fmt.Errorf("transaction not included in block %v", blk.Block.Height)
-	}
-
-	var predecessors []*evmtypes.MsgEthereumTx
-	for _, txBz := range blk.Block.Txs[:transaction.Index] {
-		tx, err := a.clientCtx.TxConfig.TxDecoder()(txBz)
-		if err != nil {
-			a.logger.Debug("failed to decode transaction in block", "height", blk.Block.Height, "error", err.Error())
-			continue
-		}
-		for _, msg := range tx.GetMsgs() {
-			ethMsg, ok := msg.(*evmtypes.MsgEthereumTx)
-			if !ok {
-				continue
-			}
-
-			predecessors = append(predecessors, ethMsg)
-		}
-	}
-
-	tx, err := a.clientCtx.TxConfig.TxDecoder()(transaction.Tx)
-	if err != nil {
-		a.logger.Debug("tx not found", "hash", hash)
-		return nil, err
-	}
-
-	// add predecessor messages in current cosmos tx
-	for i := 0; i < parsedTx.MsgIndex; i++ {
-		ethMsg, ok := tx.GetMsgs()[i].(*evmtypes.MsgEthereumTx)
-		if !ok {
-			continue
-		}
-		predecessors = append(predecessors, ethMsg)
-	}
-
-	ethMessage, ok := tx.GetMsgs()[parsedTx.MsgIndex].(*evmtypes.MsgEthereumTx)
-	if !ok {
-		a.logger.Debug("invalid transaction type", "type", fmt.Sprintf("%T", tx))
-		return nil, fmt.Errorf("invalid transaction type %T", tx)
-	}
-
-	traceTxRequest := evmtypes.QueryTraceTxRequest{
-		Msg:          ethMessage,
-		Predecessors: predecessors,
-		BlockNumber:  blk.Block.Height,
-		BlockTime:    blk.Block.Time,
-		BlockHash:    common.Bytes2Hex(blk.BlockID.Hash),
-	}
-
-	if config != nil {
-		traceTxRequest.TraceConfig = config
-	}
-
-	// minus one to get the context of block beginning
-	contextHeight := transaction.Height - 1
-	if contextHeight < 1 {
-		// 0 is a special value in `ContextWithHeight`
-		contextHeight = 1
-	}
-	traceResult, err := a.queryClient.TraceTx(rpctypes.ContextWithHeight(contextHeight), &traceTxRequest)
-	if err != nil {
-		return nil, err
-	}
-
-	// Response format is unknown due to custom tracer config param
-	// More information can be found here https://geth.ethereum.org/docs/dapp/tracing-filtered
-	var decodedResult interface{}
-	err = json.Unmarshal(traceResult.Data, &decodedResult)
-	if err != nil {
-		return nil, err
-	}
-
-	return decodedResult, nil
+	return a.backend.TraceTransaction(hash, config)
 }
 
 // TraceBlockByNumber returns the structured logs created during the execution of
@@ -190,7 +80,7 @@ func (a *API) TraceBlockByNumber(height rpctypes.BlockNumber, config *evmtypes.T
 		return nil, err
 	}
 
-	return a.traceBlock(height, config, resBlock)
+	return a.backend.TraceBlock(rpctypes.BlockNumber(resBlock.Block.Height), config, resBlock)
 }
 
 // TraceBlockByHash returns the structured logs created during the execution of
@@ -209,68 +99,7 @@ func (a *API) TraceBlockByHash(hash common.Hash, config *evmtypes.TraceConfig) (
 		return nil, errors.New("block not found")
 	}
 
-	return a.traceBlock(rpctypes.BlockNumber(resBlock.Block.Height), config, resBlock)
-}
-
-// traceBlock configures a new tracer according to the provided configuration, and
-// executes all the transactions contained within. The return value will be one item
-// per transaction, dependent on the requested tracer.
-func (a *API) traceBlock(height rpctypes.BlockNumber, config *evmtypes.TraceConfig, block *tmrpctypes.ResultBlock) ([]*evmtypes.TxTraceResult, error) {
-	txs := block.Block.Txs
-	txsLength := len(txs)
-
-	if txsLength == 0 {
-		// If there are no transactions return empty array
-		return []*evmtypes.TxTraceResult{}, nil
-	}
-
-	txDecoder := a.clientCtx.TxConfig.TxDecoder()
-
-	var txsMessages []*evmtypes.MsgEthereumTx
-	for i, tx := range txs {
-		decodedTx, err := txDecoder(tx)
-		if err != nil {
-			a.logger.Error("failed to decode transaction", "hash", txs[i].Hash(), "error", err.Error())
-			continue
-		}
-
-		for _, msg := range decodedTx.GetMsgs() {
-			ethMessage, ok := msg.(*evmtypes.MsgEthereumTx)
-			if !ok {
-				// Just considers Ethereum transactions
-				continue
-			}
-			txsMessages = append(txsMessages, ethMessage)
-		}
-	}
-
-	// minus one to get the context at the beginning of the block
-	contextHeight := height - 1
-	if contextHeight < 1 {
-		// 0 is a special value for `ContextWithHeight`.
-		contextHeight = 1
-	}
-	ctxWithHeight := rpctypes.ContextWithHeight(int64(contextHeight))
-
-	traceBlockRequest := &evmtypes.QueryTraceBlockRequest{
-		Txs:         txsMessages,
-		TraceConfig: config,
-		BlockNumber: block.Block.Height,
-		BlockTime:   block.Block.Time,
-		BlockHash:   common.Bytes2Hex(block.BlockID.Hash),
-	}
-
-	res, err := a.queryClient.TraceBlock(ctxWithHeight, traceBlockRequest)
-	if err != nil {
-		return nil, err
-	}
-
-	decodedResults := make([]*evmtypes.TxTraceResult, txsLength)
-	if err := json.Unmarshal(res.Data, &decodedResults); err != nil {
-		return nil, err
-	}
-
-	return decodedResults, nil
+	return a.backend.TraceBlock(rpctypes.BlockNumber(resBlock.Block.Height), config, resBlock)
 }
 
 // BlockProfile turns on goroutine profiling for nsec seconds and writes profile data to
